refactor(NewArticle): move the NewArticle component code to be inside editor page
