+++ src/components/index.ts
@@ -25,7 +25,6 @@ export * from '@/components/ProfileAuthor/ProfileAuthor'
 export * from '@/components/TextButton/TextButton'
 export * from '@/components/ProfileDescription/ProfileDescription'
 export * from '@/components/Anchor/Anchor'
-export * from '@/components/NewArticle/NewArticle'
 export * from '@/components/TextArea/TextArea'
 export * from '@/components/TextEditor/TextEditor'
 export * from '@/components/Alert/Alert'
