+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from '@/hooks/useGetArticle'
 export * from '@/hooks/useGetArticles'
 export * from '@/hooks/useMe'
 export * from '@/hooks/useProfile'
+export * from '@/hooks/useNewArticle'
